Kinda-sorta-ish integrate SvgSize module
Because the module is incomplete the integration is also incomplete.

For some reason, jQuery is not properly detecting the SVG width and
height attributes anymore. Will check if it happens in older releases
(that or use git bisect). Hm....

Also, there is something going on here that is making _handleImage() and
all subsequently run functions execute TWICE after the first time the
image selection is used without closing the dialog first. I have NO clue
what could be going on to cause that....